Create MarkdowHeading
Define a function heading() that turns a string into a markdown heading. By default, it will be a first-level heading, but there should be a chance to change it to the desired value. However, if the specified level is less than 1, return a first-level heading. Similarly, if the level is greater than 6, think of it as a sixth-level heading.